refactor: Moved meta data logs to separate directory
