+++ .gitignore
@@ -39,8 +39,8 @@ app.*.map.json
 /android/app/release
 
 # Freezed generated files
-/lib/src/models/*.freezed.dart
-/lib/src/models/*.g.dart
+/lib/src/models/**/*.freezed.dart
+/lib/src/models/**/*.g.dart
 
 # Locale generated files
 /lib/l10n/locale_keys.g.dart
