+++ test/recipe_scripts_tests/script_test_helper.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart' show fail;
 import 'package:ingredient_collector/src/models/ingredient.dart';
 import 'package:ingredient_collector/src/models/ingredient_parsing_result.dart';
-import 'package:ingredient_collector/src/models/meta_data_log.dart';
+import 'package:ingredient_collector/src/models/meta_data_logs/meta_data_log.dart';
 import 'package:ingredient_collector/src/models/recipe.dart';
 import 'package:ingredient_collector/src/models/recipe_parsing_job.dart';
 import 'package:ingredient_collector/src/models/recipe_parsing_result.dart';
@@ -45,8 +45,8 @@ Future<void> testParsingRecipes(
 
   var notWorkingUrls = <String>[];
   for (var job in jobs) {
-    var result = await collectRecipes([job], job.language ?? "de")
-        .then((value) => value.first);
+    var result =
+        await collectRecipes([job], job.language).then((value) => value.first);
     if (hasRecipeParsingErrors(result) || result.recipe == null) {
       notWorkingUrls.add(job.url.toString());
     }
